+++ api/go.mod
@@ -1,8 +1,13 @@
 module github.com/MagnusChase03/CS4389-Project
 
 replace github.com/MagnusChase03/CS4389-Project/models => ./models
+
 replace github.com/MagnusChase03/CS4389-Project/routes => ./routes
+replace github.com/MagnusChase03/CS4389-Project/routes/authRoutes => ./routes/authRoutes
+
 replace github.com/MagnusChase03/CS4389-Project/controllers => ./controllers
+replace github.com/MagnusChase03/CS4389-Project/controllers/authControllers => ./controllers/authControllers
+
 replace github.com/MagnusChase03/CS4389-Project/middleware => ./middleware
 replace github.com/MagnusChase03/CS4389-Project/utils => ./utils
 replace github.com/MagnusChase03/CS4389-Project/db => ./db
